Move teacher
