+++ parlai/tasks/convai2/agents.py
@@ -4,8 +4,7 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-from parlai.core.agents import create_task_agent_from_taskname
-from parlai.core.teachers import FbDialogTeacher
+from parlai.core.teachers import FbDialogTeacher, create_task_agent_from_taskname
 from parlai.utils.misc import warn_once
 from .build import build
