+++ parlai/tasks/fvqa/agents.py
@@ -4,9 +4,9 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-from parlai.core.agents import Teacher
 from parlai.core.image_featurizers import ImageLoader
 from parlai.core.metrics import Metrics
+from parlai.core.teachers import Teacher
 from .build import build
 
 import json
