+++ parlai/tasks/self_feeding/agents.py
@@ -10,11 +10,9 @@ import re
 
 import numpy as np
 
-from .build import build
-
-import parlai.core.agents as core_agents
-from parlai.core.teachers import ParlAIDialogTeacher
+from parlai.core.teachers import ParlAIDialogTeacher, MultiTaskTeacher
 from projects.self_feeding.utils import add_person_tokens
+from .build import build
 
 
 def _path(opt, filename, add_suffix=False):
@@ -202,7 +200,7 @@ class SelfFeedingTeacher(ParlAIDialogTeacher):
                 self.episodes.append([episode])
 
 
-class SelfFeedingMTLTeacher(core_agents.MultiTaskTeacher):
+class SelfFeedingMTLTeacher(MultiTaskTeacher):
     """
     Creates a teacher that is actually a set of teachers each based on a task string--
     each of these teachers will get called in turn, either randomly or in order. They
