+++ parlai/tasks/taskntalk/agents.py
@@ -4,7 +4,7 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-from parlai.core.agents import Teacher
+from parlai.core.teachers import Teacher
 from .build import build
 
 import json
